// apps/vanilla/fun-chat/reely-chat/src/models/request-type.enum.ts
export const enum RequestType {
  UserLogin = 'USER_LOGIN',
  UserExternalLogin = 'USER_EXTERNAL_LOGIN',
  UserLogout = 'USER_LOGOUT',
  UserExternalLogout = 'USER_EXTERNAL_LOGOUT',
  UserActive = 'USER_ACTIVE',
  UserInactive = 'USER_INACTIVE',
  MessageSend = 'MSG_SEND',
  MessageFromUser = 'MSG_FROM_USER',
  MessageRead = 'MSG_READ',
  MessageDeliver = 'MSG_DELIVER',
  MessageDelete = 'MSG_DELETE',
  MessageEdit = 'MSG_EDIT',
}